requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.1.2
django-jazzmin==2.6.0
django-import-export==3.0.1
psycopg2==2.9.3
python-dateutil==2.8.2
six==1.16.0
sqlparse==0.4.3
XlsxWriter==3.0.3
gunicorn==20.1.0
django-jazzmin==2.5.0
Pillow==9.3.0
django-crispy-forms==1.14.0
django-widget-tweaks==1.4.12